pyproject.toml
 [project]
name = "salaries-and-promotions"
version = "0.1.0"
description = ""
authors = [
    {name = "Vadim Tregulov"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.116.0,<0.117.0)",
    "uvicorn[standart] (>=0.35.0,<0.36.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "pyjwt[crypto] (>=2.10.1,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
